+++ libs/ai/src/ai/__init__.py
@@ -1,4 +1,5 @@
-from ai.metadata_generator import GeneratedMetadata, generate_metadata
+from ai.metadata_generator import generate_metadata
+from ai.metadata_generator_models import GeneratedMetadata
 from ai.providers import Provider, get_llm
 
 __all__ = ["GeneratedMetadata", "generate_metadata", "get_llm", "Provider"]
