+++ libs/ai/src/ai/metadata_generator.py
@@ -7,8 +7,8 @@ from typing import Any
 from langchain_core.language_models.chat_models import BaseChatModel
 from langchain_core.output_parsers import PydanticOutputParser
 from langchain_core.prompts import ChatPromptTemplate
-from pydantic import BaseModel, Field
 
+from ai.metadata_generator_models import GeneratedMetadata
 from ai.utils import load_prompt
 
 logger = logging.getLogger(__name__)
@@ -23,74 +23,13 @@ def _format_sample(sample_rows: list[dict[str, object]] | None) -> str:
     return "\n".join(lines)
 
 
-class AttributeInfo(BaseModel):
-    """Description of a single dataset attribute/column."""
-
-    name: str = Field(description="Column name")
-    type: str = Field(description="ISO 19110 type (e.g. integer, string (50), GM_Point)")
-    description: str = Field(description="Human-readable description of the column in French")
-
-
-class TemporalExtent(BaseModel):
-    """Temporal extent of the dataset, either a single instant or a period."""
-
-    type: str = Field(
-        description=(
-            "Either 'instant' (single date/time) or 'period' (date range). "
-            "Use 'unknown' if no temporal information can be inferred."
-        )
-    )
-    begin: str | None = Field(
-        default=None,
-        description=(
-            "Start date/time in ISO 8601 format (e.g. '2023-01-01' or '2023-01-01T00:00:00'). "
-            "Required when type='period', optional when type='instant'."
-        ),
-    )
-    end: str | None = Field(
-        default=None,
-        description=(
-            "End date/time in ISO 8601 format. Required when type='period'. "
-            "Use null if open-ended or unknown."
-        ),
-    )
-    instant: str | None = Field(
-        default=None,
-        description="Date/time in ISO 8601 format when type='instant'.",
-    )
-
-
-class GeneratedMetadata(BaseModel):
-    """Structured metadata generated by the LLM."""
-
-    title: str = Field(description="A concise, human-readable title for the dataset")
-    abstract: str = Field(description="A clear description of the dataset contents and purpose")
-    keywords: list[str] = Field(description="Relevant thematic keywords for the dataset")
-    topic_categories: list[str] = Field(
-        description=(
-            "One or more ISO 19115 topic category codes that best describe the dataset "
-            "(e.g. ['transportation', 'planningCadastre'])"
-        )
-    )
-    attribute_descriptions: list[AttributeInfo] = Field(
-        description="Description of each attribute/column of the dataset"
-    )
-    temporal_extent: TemporalExtent | None = Field(
-        default=None,
-        description=(
-            "Temporal extent of the dataset inferred from column names, sample data, or table name. "
-            "Set to null if no temporal information can be found."
-        ),
-    )
-
-
 def generate_metadata(
     table_name: str,
     column_names: list[str],
     llm: BaseChatModel,
     column_types: dict[str, str] | None = None,
     title: str | None = None,
-    extra_context: dict[str, Any] | None = None,
+    extra_context: str | dict[str, Any] | None = None,
     sample_rows: list[dict[str, object]] | None = None,
     bbox: str | None = None,
     keywords: list[str] | None = None,
@@ -110,7 +49,7 @@ def generate_metadata(
         llm: LangChain chat model instance to use
         column_types: Mapping of column name → SQL type string (optional)
         title: Human-readable title already chosen by the user (optional)
-        extra_context: Any additional context to pass to the prompt (optional)
+        extra_context: Additional context as string or dict to pass to the prompt (optional)
         sample_rows: Up to 5 sample rows from the table for richer inference (optional)
         bbox: Bounding box string "minx, miny, maxx, maxy" for geographic context (optional)
         keywords: Preferred keywords to use when relevant (optional)
@@ -168,13 +107,6 @@ def generate_metadata(
                 else "REGENERATE — For title only, minor rewording allowed to integrate current_abstract location if provided. Other fields: no reformulation."
             ),
         }
-    )
-
-    logger.info(
-        f"Generated metadata for '{table_name}': "
-        f"title={result.title}, "
-        f"abstract={result.abstract}, "
-        f"keywords={result.keywords}, topics={result.topic_categories}"
     )
 
     return result
