+++ libs/ai/src/ai/metadata_generator_models.py
@@ -1,0 +1,62 @@
+from pydantic import BaseModel, Field
+
+
+class AttributeInfo(BaseModel):
+    """Description of a single dataset attribute/column."""
+
+    name: str = Field(description="Column name")
+    type: str = Field(description="ISO 19110 type (e.g. integer, string (50), GM_Point)")
+    description: str = Field(description="Human-readable description of the column in French")
+
+
+class TemporalExtent(BaseModel):
+    """Temporal extent of the dataset, either a single instant or a period."""
+
+    type: str = Field(
+        description=(
+            "Either 'instant' (single date/time) or 'period' (date range). "
+            "Use 'unknown' if no temporal information can be inferred."
+        )
+    )
+    begin: str | None = Field(
+        default=None,
+        description=(
+            "Start date/time in ISO 8601 format (e.g. '2023-01-01' or '2023-01-01T00:00:00'). "
+            "Required when type='period', optional when type='instant'."
+        ),
+    )
+    end: str | None = Field(
+        default=None,
+        description=(
+            "End date/time in ISO 8601 format. Required when type='period'. "
+            "Use null if open-ended or unknown."
+        ),
+    )
+    instant: str | None = Field(
+        default=None,
+        description="Date/time in ISO 8601 format when type='instant'.",
+    )
+
+
+class GeneratedMetadata(BaseModel):
+    """Structured metadata generated by the LLM."""
+
+    title: str = Field(description="A concise, human-readable title for the dataset")
+    abstract: str = Field(description="A clear description of the dataset contents and purpose")
+    keywords: list[str] = Field(description="Relevant thematic keywords for the dataset")
+    topic_categories: list[str] = Field(
+        description=(
+            "One or more ISO 19115 topic category codes that best describe the dataset "
+            "(e.g. ['transportation', 'planningCadastre'])"
+        )
+    )
+    attribute_descriptions: list[AttributeInfo] | None = Field(
+        description="Description of each attribute/column of the dataset"
+    )
+    temporal_extent: TemporalExtent | None = Field(
+        default=None,
+        description=(
+            "Temporal extent of the dataset inferred from column names, sample data, or table name. "
+            "Set to null if no temporal information can be found."
+        ),
+    )
